pipewire launcjer disable for cinnnamon because of conflict with csd

+++ media/video/pipewire/pipewire-launcher.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+# we must exit if cinnamon desktop detected
+if [ "$XDG_CURRENT_DESKTOP" == "X-Cinnamon" ]:
+    exit 0
+fi
+
 # We need to kill any existing pipewire instance to restore sound
 pkill -u "${USER}" -fx /usr/bin/pipewire-pulse 1>/dev/null 2>&1
 pkill -u "${USER}" -fx /usr/bin/pipewire-media-session 1>/dev/null 2>&1
